# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: SensiLog API
  version: 1.0.0
  description: VALORANT感度・設定管理API
  contact:
    name: SensiLog Team
    url: https://sensilog.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sensilog.com
    description: Production
  - url: http://localhost:3001
    description: Development
...
paths:
...
  /analytics/performance:
    get:
      summary: パフォーマンス分析データ取得
      tags: [Analytics]
      security:
        - bearerAuth: []
      parameters:
        - name: metric
          in: query
          required: true
          schema:
            type: string
            enum: [combatScore, headshotPercentage, kdRatio, adr]
          description: 分析対象メトリクス
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
          description: グループ化単位
      responses:
        200:
          description: パフォーマンス分析データ
          content:
            application/json:
              schema:
                type: object
                properties:
                  dataPoints:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnalyticsDataPoint'
                  summary:
                    $ref: '#/components/schemas/PerformanceSummary'
                  settingsChanges:
                    type: array
                    items:
                      $ref: '#/components/schemas/SettingsChange'
  /analytics/comparison:
    post:
      summary: 期間比較分析
      tags: [Analytics]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                period1:
                  $ref: '#/components/schemas/DateRange'
                period2:
                  $ref: '#/components/schemas/DateRange'
                metrics:
                  type: array
                  items:
                    type: string
                    enum: [combatScore, headshotPercentage, kdRatio, adr]
              required:
                - period1
                - period2
                - metrics
      responses:
        200:
          description: 比較分析結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonResult'
components:
  schemas:
...
    AnalyticsDataPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        value:
          type: number
          format: float
        matchCount:
          type: integer
          description: 該当期間の試合数
      required:
        - date
        - value
    PerformanceSummary:
      type: object
      properties:
        average:
          type: number
          format: float
        best:
          type: number
          format: float
        worst:
          type: number
          format: float
        trend:
          type: string
          enum: [improving, declining, stable]
        changePercent:
          type: number
          format: float
          description: 変化率（%）
    SettingsChange:
      type: object
      properties:
        date:
          type: string
          format: date-time
        changes:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              oldValue:
                type: string
              newValue:
                type: string
    DateRange:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
        - startDate
        - endDate
    ComparisonResult:
      type: object
      properties:
        period1:
          type: object
          properties:
            label:
              type: string
            stats:
              $ref: '#/components/schemas/PerformanceSummary'
        period2:
          type: object
          properties:
            label:
              type: string
            stats:
              $ref: '#/components/schemas/PerformanceSummary'
        comparison:
          type: object
          additionalProperties:
            type: object
            properties:
              difference:
                type: number
                format: float
              percentChange:
                type: number
                format: float
              trend:
                type: string
                enum: [up, down, same]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT